feat: seperate validation file; comment method in model for code coverage; change route for new file;
